make creds exporter for jtr format
